removed code out of __init__.py
